requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
boto3==1.35.74
botocore==1.35.74
crispy-bootstrap5==2024.10
Django==4.2.16
django-bootstrap-v5==1.0.11
django-crispy-forms==2.3
django-storages==1.14.4
jmespath==1.0.1
pillow==11.0.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.4
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
